+++ src/app/modules/marketplace/components/module-train/general-conf-form/general-conf-form.component.spec.ts
@@ -13,13 +13,50 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { SidenavComponent } from '@app/layout/sidenav/sidenav.component';
 import { ModuleGeneralConfiguration } from '@app/shared/interfaces/module.interface';
 import { MediaMatcher } from '@angular/cdk/layout';
-import { OAuthStorage } from 'angular-oauth2-oidc';
+import { provideHttpClient } from '@angular/common/http';
+import { AuthService } from '@app/core/services/auth/auth.service';
+import { of } from 'rxjs';
 
 const mockDefaultFormValues: ModuleGeneralConfiguration = {
     title: { name: '', value: '', description: '' },
     docker_image: { name: '', value: '', description: '' },
     docker_tag: { name: '', value: '', description: '' },
     service: { name: '', value: '', description: '' },
+};
+
+const mockedProfile = {
+    info: {
+        exp: 1693908513,
+        iat: 1693907913,
+        auth_time: 1693907911,
+        jti: '00000000-c9e1-44b7-b313-4bde8fba70fa',
+        iss: 'https://aai-demo.egi.eu/auth/realms/egi',
+        aud: 'ai4eosc-dashboard',
+        sub: 'test@egi.eu',
+        typ: 'ID',
+        azp: 'ai4eosc-dashboard',
+        nonce: 'WnVHR3ZpOVoyVlFwcjVGTEtIRWhyUTZ0eXJYVHZxN1M4TX5MRzVKWVJYVHZx',
+        session_state: '00000000-818c-46d4-ad87-1b9a1c22c43f',
+        at_hash: 'gdEA9VsgdEA9V-mubWhBWw',
+        sid: 'b27a9b7a-818c-46d4-ad87-1b9a1818c43f',
+        voperson_verified_email: ['test@ifca.unican.es'],
+        email_verified: true,
+        name: 'AI4EOSC Dasboard Test',
+        preferred_username: 'ai4dash',
+        eduperson_assurance: [
+            'https://refeds.org/assurance/IAP/low',
+            'https://aai.egi.eu/LoA#Low',
+        ],
+        given_name: 'AI4EOSC Dasboard ',
+        family_name: 'Test',
+        email: 'test@ifca.unican.es',
+        eduperson_entitlement: [
+            'urn:mace:egi.eu:group:vo.ai4eosc.eu:role=member#aai.egi.eu',
+            'urn:mace:egi.eu:group:vo.ai4eosc.eu:role=vm_operator#aai.egi.eu',
+            'urn:mace:egi.eu:group:vo.imagine-ai.eu:role=member#aai.egi.eu',
+            'urn:mace:egi.eu:group:vo.imagine-ai.eu:role=vm_operator#aai.egi.eu',
+        ],
+    },
 };
 
 const mockedMediaQueryList: MediaQueryList = {
@@ -35,6 +72,14 @@ const mockedMediaQueryList: MediaQueryList = {
 
 const mockedMediaMatcher: any = {
     matchMedia: jest.fn().mockReturnValue(mockedMediaQueryList),
+};
+
+const mockedAuthService: any = {
+    isAuthenticated: jest.fn(),
+    userProfileSubject: of({}),
+    loadUserProfile: jest.fn().mockReturnValue(Promise.resolve(mockedProfile)),
+    login: jest.fn(),
+    logout: jest.fn(),
 };
 
 describe('GeneralConfFormComponent', () => {
@@ -56,9 +101,10 @@ describe('GeneralConfFormComponent', () => {
                 NoopAnimationsModule,
             ],
             providers: [
+                provideHttpClient(),
                 FormGroupDirective,
                 FormBuilder,
-                { provide: OAuthStorage, useValue: {} },
+                { provide: AuthService, useValue: mockedAuthService },
                 { provide: FormGroupDirective, useValue: formGroupDirective },
                 { provide: MediaMatcher, useValue: mockedMediaMatcher },
             ],