+++ src/app/modules/marketplace/components/module-train/stepper-form/stepper-form.component.ts
@@ -15,7 +15,7 @@ import { FormBuilder, FormGroup } from '@angular/forms';
 import { MatSlideToggleChange } from '@angular/material/slide-toggle';
 import { Router } from '@angular/router';
 import { DeploymentsService } from '@app/modules/deployments/services/deployments-service/deployments.service';
-import { statusReturn } from '@app/shared/interfaces/deployment.interface';
+import { StatusReturn } from '@app/shared/interfaces/deployment.interface';
 import { TrainModuleRequest } from '@app/shared/interfaces/module.interface';
 import { SnackbarService } from '@app/shared/services/snackbar/snackbar.service';
 import { Observable } from 'rxjs';
@@ -46,15 +46,16 @@ export class StepperFormComponent implements OnInit {
     }
 
     @Input() title!: string;
+    @Input() numberOfSteps!: number;
     @Input() step1!: TemplateRef<unknown>;
     @Input() step2!: TemplateRef<unknown>;
     @Input() step3!: TemplateRef<unknown>;
     @Input() step1Form!: FormGroup;
     @Input() step2Form!: FormGroup;
-    @Input() step3Form!: FormGroup;
+    @Input() step3Form?: FormGroup;
     @Input() step1Title!: string;
     @Input() step2Title!: string;
-    @Input() step3Title!: string;
+    @Input() step3Title?: string;
     @Input() isLoading!: boolean;
 
     @Output() showHelpButtonEvent = new EventEmitter<MatSlideToggleChange>();
@@ -84,8 +85,7 @@ export class StepperFormComponent implements OnInit {
 
     submitTrainingRequest() {
         this.isLoading = true;
-
-        let request: Observable<statusReturn>;
+        let request: Observable<StatusReturn>;
         const data: TrainModuleRequest = {
             general: {
                 title:
@@ -104,81 +104,116 @@ export class StepperFormComponent implements OnInit {
                 jupyter_password:
                     this.step1Form.getRawValue().generalConfForm
                         .serviceToRunPassInput,
+                cvat_username:
+                    this.step1Form.getRawValue().generalConfForm
+                        .cvatUsernameInput,
+                cvat_password:
+                    this.step1Form.getRawValue().generalConfForm
+                        .cvatPasswordInput,
             },
-            hardware: {
+        };
+
+        if (this.title == 'CVAT Image Annotation') {
+            data.storage = {
+                rclone_conf:
+                    this.step2Form.value.storageConfForm.rcloneConfInput,
+                rclone_url:
+                    this.step2Form.value.storageConfForm.storageUrlInput,
+                rclone_vendor:
+                    this.step2Form.value.storageConfForm.rcloneVendorSelect,
+                rclone_user:
+                    this.step2Form.value.storageConfForm.rcloneUserInput,
+                rclone_password:
+                    this.step2Form.value.storageConfForm.rclonePasswordInput,
+                cvat_backup:
+                    this.step2Form.value.storageConfForm.snapshotDatasetSelect,
+            };
+            request = this.deploymentsService.trainTool('ai4os-cvat', data);
+        } else {
+            data.hardware = {
                 cpu_num: this.step2Form.value.hardwareConfForm.cpuNumberInput,
                 ram: this.step2Form.value.hardwareConfForm.ramMemoryInput,
                 disk: this.step2Form.value.hardwareConfForm.diskMemoryInput,
                 gpu_num: this.step2Form.value.hardwareConfForm.gpuNumberInput,
                 gpu_type: this.step2Form.value.hardwareConfForm.gpuModelSelect,
-            },
-        };
+            };
 
-        if (this.title == 'Federated learning server') {
-            data.configuration = {
-                rounds: this.step3Form.value.federatedConfForm.roundsInput,
-                metric: this.step3Form.value.federatedConfForm.metricInput,
-                min_fit_clients:
-                    this.step3Form.value.federatedConfForm.minFitClientsInput,
-                min_available_clients:
-                    this.step3Form.value.federatedConfForm
-                        .minAvailableClientsInput,
-                strategy:
-                    this.step3Form.value.federatedConfForm
-                        .strategyOptionsSelect,
-                mu:
-                    this.step3Form.value.federatedConfForm
-                        .strategyOptionsSelect === 'FedProx strategy (FedProx)'
-                        ? this.step3Form.value.federatedConfForm.muInput
+            if (this.title == 'Federated learning server') {
+                data.configuration = {
+                    rounds: this.step3Form!.value.federatedConfForm.roundsInput,
+                    metric: this.step3Form!.value.federatedConfForm.metricInput,
+                    min_fit_clients:
+                        this.step3Form!.value.federatedConfForm
+                            .minFitClientsInput,
+                    min_available_clients:
+                        this.step3Form!.value.federatedConfForm
+                            .minAvailableClientsInput,
+                    strategy:
+                        this.step3Form!.value.federatedConfForm
+                            .strategyOptionsSelect,
+                    mu:
+                        this.step3Form!.value.federatedConfForm
+                            .strategyOptionsSelect ===
+                        'FedProx strategy (FedProx)'
+                            ? this.step3Form!.value.federatedConfForm.muInput
+                            : null,
+                    fl:
+                        this.step3Form!.value.federatedConfForm
+                            .strategyOptionsSelect ===
+                        'Federated Averaging with Momentum (FedAvgM)'
+                            ? this.step3Form!.value.federatedConfForm.flInput
+                            : null,
+                    momentum:
+                        this.step3Form!.value.federatedConfForm
+                            .strategyOptionsSelect ===
+                        'Federated Averaging with Momentum (FedAvgM)'
+                            ? this.step3Form!.value.federatedConfForm
+                                .momentumInput
+                            : null,
+                    dp: this.step3Form!.value.federatedConfForm.dpInput,
+                    noise_mult: this.step3Form!.value.federatedConfForm.dpInput
+                        ? this.step3Form!.value.federatedConfForm.noiseMultInput
                         : null,
-                fl:
-                    this.step3Form.value.federatedConfForm
-                        .strategyOptionsSelect ===
-                    'Federated Averaging with Momentum (FedAvgM)'
-                        ? this.step3Form.value.federatedConfForm.flInput
+                    sampled_clients: this.step3Form!.value.federatedConfForm
+                        .dpInput
+                        ? this.step3Form!.value.federatedConfForm
+                            .sampledClientsNumInput
                         : null,
-                momentum:
-                    this.step3Form.value.federatedConfForm
-                        .strategyOptionsSelect ===
-                    'Federated Averaging with Momentum (FedAvgM)'
-                        ? this.step3Form.value.federatedConfForm.momentumInput
+                    clip_norm: this.step3Form!.value.federatedConfForm.dpInput
+                        ? this.step3Form!.value.federatedConfForm
+                            .clippingNormInput
                         : null,
-                dp: this.step3Form.value.federatedConfForm.dpInput,
-                noise_mult: this.step3Form.value.federatedConfForm.dpInput
-                    ? this.step3Form.value.federatedConfForm.noiseMultInput
-                    : null,
-                sampled_clients: this.step3Form.value.federatedConfForm.dpInput
-                    ? this.step3Form.value.federatedConfForm
-                        .sampledClientsNumInput
-                    : null,
-                clip_norm: this.step3Form.value.federatedConfForm.dpInput
-                    ? this.step3Form.value.federatedConfForm.clippingNormInput
-                    : null,
-            };
-            request = this.deploymentsService.trainTool(data);
-        } else {
-            data.storage = {
-                rclone_conf:
-                    this.step3Form.value.storageConfForm.rcloneConfInput,
-                rclone_url:
-                    this.step3Form.value.storageConfForm.storageUrlInput,
-                rclone_vendor:
-                    this.step3Form.value.storageConfForm.rcloneVendorSelect,
-                rclone_user:
-                    this.step3Form.value.storageConfForm.rcloneUserInput,
-                rclone_password:
-                    this.step3Form.value.storageConfForm.rclonePasswordInput,
-                datasets:
-                    this.step3Form.value.storageConfForm.datasetsList[0]
-                        ?.doi === ''
-                        ? []
-                        : this.step3Form.value.storageConfForm.datasetsList,
-            };
-            request = this.deploymentsService.postTrainModule(data);
+                };
+                request = this.deploymentsService.trainTool(
+                    'ai4os-federated-server',
+                    data
+                );
+            } else {
+                data.storage = {
+                    rclone_conf:
+                        this.step3Form!.value.storageConfForm.rcloneConfInput,
+                    rclone_url:
+                        this.step3Form!.value.storageConfForm.storageUrlInput,
+                    rclone_vendor:
+                        this.step3Form!.value.storageConfForm
+                            .rcloneVendorSelect,
+                    rclone_user:
+                        this.step3Form!.value.storageConfForm.rcloneUserInput,
+                    rclone_password:
+                        this.step3Form!.value.storageConfForm
+                            .rclonePasswordInput,
+                    datasets:
+                        this.step3Form!.value.storageConfForm.datasetsList[0]
+                            ?.doi === ''
+                            ? []
+                            : this.step3Form!.value.storageConfForm
+                                .datasetsList,
+                };
+                request = this.deploymentsService.postTrainModule(data);
+            }
         }
-
         request.subscribe({
-            next: (result: statusReturn) => {
+            next: (result: StatusReturn) => {
                 this.isLoading = false;
 
                 if (result && result.status == 'success') {