+++ src/app/modules/marketplace/services/tools-service/tools.service.ts
@@ -3,6 +3,7 @@ import { Injectable } from '@angular/core';
 import { AppConfigService } from '@app/core/services/app-config/app-config.service';
 import { TagObject } from '@app/data/types/tags';
 import {
+    CvatToolConfiguration,
     FederatedServerToolConfiguration,
     Module,
     ModuleSummary,
@@ -42,13 +43,23 @@ export class ToolsService {
     }
 
     getFederatedServerConfiguration(
-        moduleName: string
+        toolName: string
     ): Observable<FederatedServerToolConfiguration> {
         const url = `${base}${endpoints.toolConfiguration.replace(
             ':name',
-            moduleName
+            toolName
         )}`;
         return this.http.get<FederatedServerToolConfiguration>(url, {
+            params: this.voParam,
+        });
+    }
+
+    getCvatConfiguration(toolName: string): Observable<CvatToolConfiguration> {
+        const url = `${base}${endpoints.toolConfiguration.replace(
+            ':name',
+            toolName
+        )}`;
+        return this.http.get<CvatToolConfiguration>(url, {
             params: this.voParam,
         });
     }